Sprint 4 (#591 E5e): split AdminAssetsManager — extract per-screen conditional assets
Moves the per-screen conditional enqueue logic into a new AdminConditionalAssets
class (enqueue_conditional_assets + the certificates-dashboard / submissions-bulk
/ move-submissions / submission-edit enqueues and their is_*_page guards +
resolve_single_filter_form_id + is_documentation_tab). AdminAssetsManager
(899 → 511) keeps the orchestrator (enqueue_admin_assets), the always-on core
assets (core module, CSS, JS modules, dark-mode, code-editor) and localization.
enqueue_admin_assets delegates the conditional part to the new class.

is_settings_page changed private → public static so the moved code can call it;
the three public entry points are unchanged. No asset handle, path, dep, version,
localized-data, hook, or screen condition changed. New class carries @covers +
setUp preload (verified 110/239). Behavior-preserving — AdminAssetsManager suite
green (22 tests); PHPStan L8 + WPCS clean.
